+++ rundeckapp/src/test/groovy/rundeck/ScheduledExecutionServiceTests.groovy
@@ -16,22 +16,14 @@ package rundeck
  */
 
 //import grails.test.GrailsUnitTestCase
-import grails.test.mixin.Mock
-import grails.test.mixin.TestFor
+
 import groovy.mock.interceptor.MockFor
-import org.junit.Ignore
-import rundeck.Execution
-import rundeck.Notification
-import rundeck.Option
-import rundeck.JobExec
-import rundeck.PluginStep
-import rundeck.ScheduledExecution
-import rundeck.CommandExec
-import rundeck.Workflow
-import rundeck.WorkflowStep
+import org.junit.Test
+import rundeck.*
 import rundeck.controllers.ScheduledExecutionController
 import rundeck.services.FrameworkService
 import rundeck.services.ScheduledExecutionService
+
 import static org.junit.Assert.*
 
 /*
@@ -41,11 +33,6 @@ import static org.junit.Assert.*
 * Created: Jul 29, 2010 4:38:24 PM
 * $Id$
 */
-/********
- * NEEDS to be changed to Spec
- *******/@Ignore
-@TestFor(ScheduledExecutionService)
-@Mock([ScheduledExecution,Workflow,WorkflowStep,CommandExec,JobExec,Execution, Option,Notification])
 public class ScheduledExecutionServiceTests {
 
 
@@ -54,6 +41,7 @@ public class ScheduledExecutionServiceTests {
         assertNotNull(result)
         assertEquals(expected, result)
     }
+    @Test
     public void testParseParamNotificationsSuccess() {
         assertParseParamNotifications(
                 [[eventTrigger: ScheduledExecutionController.ONSUCCESS_TRIGGER_NAME,
@@ -63,6 +51,7 @@ public class ScheduledExecutionServiceTests {
                         (ScheduledExecutionController.NOTIFY_SUCCESS_RECIPIENTS): 'c@example.com,d@example.com']
         )
     }
+    @Test
     public void testParseParamNotificationsSuccess_subject() {
         assertParseParamNotifications(
                 [[eventTrigger: ScheduledExecutionController.ONSUCCESS_TRIGGER_NAME,
@@ -75,7 +64,7 @@ public class ScheduledExecutionServiceTests {
                         (ScheduledExecutionController.NOTIFY_SUCCESS_SUBJECT): 'blah']
         )
     }
-
+    @Test
     public void testParseParamNotificationsSuccessUrl() {
         assertParseParamNotifications(
                 [[eventTrigger: ScheduledExecutionController.ONSUCCESS_TRIGGER_NAME,
@@ -85,6 +74,7 @@ public class ScheduledExecutionServiceTests {
                         (ScheduledExecutionController.NOTIFY_SUCCESS_URL): 'http://blah.com']
         )
     }
+    @Test
     public void testParseParamNotificationsFailure() {
         assertParseParamNotifications(
                 [[eventTrigger: ScheduledExecutionController.ONFAILURE_TRIGGER_NAME,
@@ -94,6 +84,7 @@ public class ScheduledExecutionServiceTests {
                         (ScheduledExecutionController.NOTIFY_FAILURE_RECIPIENTS): 'c@example.com,d@example.com']
         )
     }
+    @Test
     public void testParseParamNotificationsFailure_subject() {
         assertParseParamNotifications(
                 [[eventTrigger: ScheduledExecutionController.ONFAILURE_TRIGGER_NAME,
@@ -105,7 +96,7 @@ public class ScheduledExecutionServiceTests {
                         (ScheduledExecutionController.NOTIFY_FAILURE_SUBJECT): 'elf']
         )
     }
-
+    @Test
     public void testParseParamNotificationsFailureUrl() {
         assertParseParamNotifications(
                 [[eventTrigger: ScheduledExecutionController.ONFAILURE_TRIGGER_NAME,
@@ -115,7 +106,7 @@ public class ScheduledExecutionServiceTests {
                         (ScheduledExecutionController.NOTIFY_FAILURE_URL): 'http://blah.com']
         )
     }
-
+    @Test
     public void testParseParamNotificationsRetryableFailure() {
         assertParseParamNotifications(
                 [[eventTrigger: ScheduledExecutionController.ONRETRYABLEFAILURE_TRIGGER_NAME,
@@ -125,6 +116,7 @@ public class ScheduledExecutionServiceTests {
                         (ScheduledExecutionController.NOTIFY_RETRYABLEFAILURE_RECIPIENTS): 'c@example.com,d@example.com']
         )
     }
+    @Test
     public void testParseParamNotificationsRetryableFailure_subject() {
         assertParseParamNotifications(
                 [[eventTrigger: ScheduledExecutionController.ONRETRYABLEFAILURE_TRIGGER_NAME,
@@ -136,7 +128,7 @@ public class ScheduledExecutionServiceTests {
                         (ScheduledExecutionController.NOTIFY_RETRYABLEFAILURE_SUBJECT): 'elf']
         )
     }
-
+    @Test
     public void testParseParamNotificationsRetryableFailureUrl() {
         assertParseParamNotifications(
                 [[eventTrigger: ScheduledExecutionController.ONRETRYABLEFAILURE_TRIGGER_NAME,
@@ -146,6 +138,7 @@ public class ScheduledExecutionServiceTests {
                         (ScheduledExecutionController.NOTIFY_RETRYABLEFAILURE_URL): 'http://blah.com']
         )
     }
+    @Test
     public void testParseParamNotificationsStart() {
         assertParseParamNotifications(
                 [[eventTrigger: ScheduledExecutionController.ONFAILURE_TRIGGER_NAME,
@@ -155,6 +148,7 @@ public class ScheduledExecutionServiceTests {
                         (ScheduledExecutionController.NOTIFY_FAILURE_RECIPIENTS): 'c@example.com,d@example.com']
         )
     }
+    @Test
     public void testParseParamNotificationsStart_subject() {
         assertParseParamNotifications(
                 [[eventTrigger: ScheduledExecutionController.ONFAILURE_TRIGGER_NAME,
@@ -167,7 +161,7 @@ public class ScheduledExecutionServiceTests {
                         (ScheduledExecutionController.NOTIFY_FAILURE_SUBJECT): 'rango']
         )
     }
-
+    @Test
     public void testParseParamNotificationsStartUrl() {
         assertParseParamNotifications(
                 [[eventTrigger: ScheduledExecutionController.ONSTART_TRIGGER_NAME,
@@ -177,6 +171,7 @@ public class ScheduledExecutionServiceTests {
                         (ScheduledExecutionController.NOTIFY_START_URL): 'http://blah.com']
         )
     }
+    @Test
     public void testParseParamNotificationsSuccessPluginEnabled() {
         assertParseParamNotifications(
                 [[eventTrigger: ScheduledExecutionController.ONSUCCESS_TRIGGER_NAME, type: 'plugin1', configuration: [:]]],
@@ -195,6 +190,7 @@ public class ScheduledExecutionServiceTests {
                 ]
         )
     }
+    @Test
     public void testParseParamNotificationsFailurePluginEnabled() {
         assertParseParamNotifications(
                 [[eventTrigger: ScheduledExecutionController.ONFAILURE_TRIGGER_NAME, type: 'plugin1', configuration: [:]]],
@@ -213,6 +209,7 @@ public class ScheduledExecutionServiceTests {
                 ]
         )
     }
+    @Test
     public void testParseParamNotificationsRetryableFailurePluginEnabled() {
         assertParseParamNotifications(
                 [[eventTrigger: ScheduledExecutionController.ONRETRYABLEFAILURE_TRIGGER_NAME, type: 'plugin1', configuration: [:]]],
@@ -231,6 +228,7 @@ public class ScheduledExecutionServiceTests {
                 ]
         )
     }
+    @Test
     public void testParseParamNotificationsStartPluginEnabled() {
         assertParseParamNotifications(
                 [[eventTrigger: ScheduledExecutionController.ONSTART_TRIGGER_NAME, type: 'plugin1', configuration: [:]]],
@@ -249,6 +247,7 @@ public class ScheduledExecutionServiceTests {
                 ]
         )
     }
+    @Test
     public void testParseParamNotificationsSuccessPluginDisabled() {
         assertParseParamNotifications(
                 [],
@@ -267,6 +266,7 @@ public class ScheduledExecutionServiceTests {
                 ]
         )
     }
+    @Test
     public void testParseParamNotificationsSuccessPluginConfiguration() {
         assertParseParamNotifications(
                 [[eventTrigger: ScheduledExecutionController.ONSUCCESS_TRIGGER_NAME, type: 'plugin1', configuration: [a:'b',c:'def']]],
@@ -285,6 +285,7 @@ public class ScheduledExecutionServiceTests {
                 ]
         )
     }
+    @Test
     public void testParseParamNotificationsSuccessPluginMultiple() {
         assertParseParamNotifications(
                 [
@@ -310,6 +311,7 @@ public class ScheduledExecutionServiceTests {
                 ]
         )
     }
+    @Test
     public void testParseParamEmailNotificationsAttachedFile() {
         assertParseParamNotifications(
                 [[eventTrigger: ScheduledExecutionController.ONSUCCESS_TRIGGER_NAME,
@@ -322,6 +324,7 @@ public class ScheduledExecutionServiceTests {
                 ]
         )
     }
+    @Test
     public void testParseParamEmailNotificationsAttachedInline() {
         assertParseParamNotifications(
                 [[eventTrigger: ScheduledExecutionController.ONSUCCESS_TRIGGER_NAME,
@@ -333,179 +336,5 @@ public class ScheduledExecutionServiceTests {
                  (ScheduledExecutionController.NOTIFY_SUCCESS_ATTACH_TYPE):'inline'
                 ]
         )
-    }
-
-    public void testGetGroups(){
-        def schedlist=[new ScheduledExecution(jobName:'test1',groupPath:'group1'),new ScheduledExecution(jobName:'test2',groupPath:null)]
-
-        ScheduledExecution.metaClass.static.findAllByProject={proj-> return schedlist}
-
-        ScheduledExecutionService test = new ScheduledExecutionService()
-        def fwkControl = new MockFor(FrameworkService, true)
-
-        fwkControl.demand.authResourceForJob{job->
-            [type:'job',name:job.jobName,group:job.groupPath?:'']
-        }
-        fwkControl.demand.authResourceForJob{job->
-            [type:'job',name:job.jobName,group:job.groupPath?:'']
-        }
-        fwkControl.demand.authorizeProjectResources{fwk,Set resset,actionset,proj->
-            assertEquals 2,resset.size()
-            def list = resset.sort{a,b->a.name<=>b.name}
-            assertEquals([type:'job',name:'test1',group:'group1'],list[0])
-            assertEquals([type:'job',name:'test2',group:''],list[1])
-            
-            assertEquals 1,actionset.size()
-            assertEquals 'read',actionset.iterator().next()
-
-            assertEquals 'proj1',proj
-
-            return [[authorized:true,resource:list[0]],[authorized:false,resource:list[1]]]
-        }
-        test.frameworkService = fwkControl.proxyInstance()
-        def result=test.getGroups("proj1",null)
-        assertEquals 1,result.size()
-        assertEquals 1,result['group1']
-
-    }
-
-    void testBlah() {
-        def (ScheduledExecution job1, String serverUUID2, ScheduledExecution job2, ScheduledExecution job3,
-        String serverUUID) = setupTestClaimScheduledJobs()
-        ScheduledExecutionService testService = new ScheduledExecutionService()
-
-        assertEquals(null, job1.serverNodeUUID)
-        assertEquals(serverUUID2, job2.serverNodeUUID)
-        assertEquals(null, job3.serverNodeUUID)
-
-        job1.serverNodeUUID=serverUUID
-        assertNotNull(job1.save(flush: true))
-
-        job1 = ScheduledExecution.get(job1.id)
-
-        assertEquals(serverUUID, job1.serverNodeUUID)
-    }
-
-    private List setupTestClaimScheduledJobs() {
-
-        def serverUUID = UUID.randomUUID().toString()
-        def serverUUID2 = UUID.randomUUID().toString()
-        ScheduledExecution job1 = new ScheduledExecution(
-                jobName: 'blue',
-                project: 'AProject',
-                groupPath: 'some/where',
-                description: 'a job',
-                argString: '-a b -c d',
-                workflow: new Workflow(keepgoing: true, commands:
-                        [new CommandExec([adhocRemoteString: 'test buddy'])]),
-                serverNodeUUID: null,
-                scheduled: true
-        )
-        assertTrue(job1.validate())
-        assertNotNull(job1.save())
-        ScheduledExecution job2 = new ScheduledExecution(
-                jobName: 'blue2',
-                project: 'AProject2',
-                groupPath: 'some/where2',
-                description: 'a job2',
-                argString: '-a b -c d2',
-                workflow: new Workflow(keepgoing: true, commands:
-                        [new CommandExec([adhocRemoteString: 'test buddy2'])]),
-                serverNodeUUID: serverUUID2,
-                scheduled: true
-        )
-        assertTrue(job2.validate())
-        assertNotNull(job2.save())
-        ScheduledExecution job3 = new ScheduledExecution(
-                jobName: 'blue2',
-                project: 'AProject2',
-                groupPath: 'some/where2',
-                description: 'a job2',
-                argString: '-a b -c d2',
-                workflow: new Workflow(keepgoing: true, commands:
-                        [new CommandExec([adhocRemoteString: 'test buddy2'])]),
-                scheduled: false,
-        )
-        assertTrue(job3.validate())
-        assertNotNull(job3.save())
-        [job1, serverUUID2, job2, job3, serverUUID]
-    }
-
-    public void testValidateWorkflow() {
-        ScheduledExecutionService testService = new ScheduledExecutionService()
-
-        def cmdExecProps = [adhocRemoteString: 'test buddy2']
-        def jobrefWorkflowStepProps = [jobName: "name", jobGroup: "group"]
-        def jobrefNodeStepProps = [jobName: "name2", jobGroup: "group2", nodeStep: true]
-        def pluginNodeStepProps = [type: 'plug1', nodeStep: true,]
-        def pluginWorkflowStepProps = [type: 'plug1', nodeStep: false]
-
-        //simple
-        assertValidateWorkflow([new CommandExec(cmdExecProps)], testService, true)
-
-        //exec step cannot have a workflow step jobref error handler
-        assertValidateWorkflow(
-                [new CommandExec(cmdExecProps + [errorHandler: new JobExec(jobrefWorkflowStepProps)])],
-                testService, false)
-
-        //exec step can have a job ref (node step) error handler
-        assertValidateWorkflow(
-                [new CommandExec(cmdExecProps + [errorHandler: new JobExec(jobrefNodeStepProps)])],
-                testService, true)
-
-        //exec step cannot have a workflow step plugin error handler
-        assertValidateWorkflow(
-                [new CommandExec(cmdExecProps + [errorHandler: new PluginStep(pluginWorkflowStepProps)])],
-                testService, false)
-
-        //exec step can have a node step plugin error handler
-        assertValidateWorkflow(
-                [new CommandExec(cmdExecProps + [errorHandler: new PluginStep(pluginNodeStepProps)])],
-                testService, true)
-
-        //node step plugin cannot have a workflow step error handler
-        assertValidateWorkflow(
-                [new PluginStep(pluginNodeStepProps + [errorHandler: new JobExec(jobrefWorkflowStepProps)])],
-                testService, false)
-
-        //workflow step plugin can have a workflow step error handler
-        assertValidateWorkflow(
-                [new PluginStep(pluginWorkflowStepProps + [ errorHandler: new JobExec(jobrefWorkflowStepProps)])],
-                testService, true)
-
-        //job ref(workflow step) can have another as error handler
-        assertValidateWorkflow(
-                [new JobExec(jobrefWorkflowStepProps + [ errorHandler: new JobExec(jobrefWorkflowStepProps)])],
-                testService, true)
-
-        //job ref(workflow step) can have a plugin workflow step handler
-        assertValidateWorkflow(
-                [new JobExec(jobrefWorkflowStepProps + [ errorHandler: new PluginStep(pluginWorkflowStepProps)])],
-                testService, true)
-
-        //job ref(workflow step) can have a node step plugin erro handler
-        assertValidateWorkflow(
-                [new JobExec(jobrefWorkflowStepProps + [ errorHandler: new PluginStep(pluginNodeStepProps)])],
-                testService, true)
-
-
-    }
-
-    private void assertValidateWorkflow(List<WorkflowStep> commands, ScheduledExecutionService testService, boolean valid) {
-        def workflow = new Workflow(keepgoing: true, commands: commands, strategy: 'node-first')
-        ScheduledExecution scheduledExecution = new ScheduledExecution(
-                jobName: 'blue2',
-                project: 'AProject2',
-                groupPath: 'some/where2',
-                description: 'a job2',
-                argString: '-a b -c d2',
-                workflow: workflow,
-                scheduled: false,
-        )
-        assert valid == testService.validateWorkflow(workflow, scheduledExecution)
-        assert !valid == scheduledExecution.hasErrors()
-        assert !valid == scheduledExecution.errors.hasFieldErrors('workflow')
-        assert !valid == workflow.commands[0].hasErrors()
-        assert !valid == workflow.commands[0].errors.hasFieldErrors('errorHandler')
     }
 }
